refactor: separate models and views within respective folders
fix booking serializer apartment existing check
