+++ accommodations/models/__init__.py
@@ -1,0 +1,3 @@
+from .accommodation import *
+from .apartment import * 
+from .hotel import *
