+++ accommodations/views.py
@@ -1,4 +1,2 @@
-from .accomodation_views import *
-from .apartement_views import *
-from .hotel_views import *
 
+from .views import *
