+++ accommodations/views/__init__.py
@@ -1,0 +1,3 @@
+from .accommodation import *
+from .apartment import *  # Replace with specific imports if needed
+from .hotel import *  # Replace with specific imports if needed
