+++ accommodations/views/apartment.py
@@ -1,7 +1,7 @@
 from rest_framework import generics
 from drf_spectacular.utils import extend_schema
-from .models import Apartment
-from .serializers import ApartmentSerializer
+from ..models import Apartment
+from ..serializers import ApartmentSerializer
 from src.helpers.pagination import IDCursorPagination
 
 # Apartments
