+++ bookings/serializers.py
@@ -36,9 +36,9 @@ class BookingSerializer(serializers.ModelSerializer):
         if start_date and end_date and end_date <= start_date:
             raise serializers.ValidationError("End date must be after start date")
         
-        apartment = Apartment.objects.get(id=data['accommodation_id'])
+        try:
+            apartment = Apartment.objects.get(id=data['accommodation_id'])
 
-        if apartment:
             overlapping = Booking(
                 accommodation=apartment,
                 start_date=data['start_date'],
@@ -48,7 +48,10 @@ class BookingSerializer(serializers.ModelSerializer):
             if overlapping:
                 raise serializers.ValidationError("Overlapping bookings are not allowed for apartments.")
         
-        return data
+            return data
+        except Apartment.DoesNotExist:
+            return data
+        
 
     def create(self, validated_data):
         """Create booking with accommodation"""
